+++ tools/test-js-with-node
@@ -91,7 +91,6 @@ EXEMPT_FILES = make_set(
         "web/src/css_variables.d.ts",
         "web/src/css_variables.js",
         "web/src/custom_profile_fields_ui.ts",
-        "web/src/dark_theme.ts",
         "web/src/debug.ts",
         "web/src/demo_organizations_ui.js",
         "web/src/deprecated_feature_notice.ts",
@@ -248,6 +247,7 @@ EXEMPT_FILES = make_set(
         "web/src/stream_ui_updates.js",
         "web/src/submessage.ts",
         "web/src/subscriber_api.ts",
+        "web/src/theme.ts",
         "web/src/timerender.ts",
         "web/src/tippyjs.ts",
         "web/src/todo_widget.js",
